Split local_3d.py by modality, remove strong model refs, soften CUDA gates (#28)
* Split local_3d.py into text_to_3d.py + image_to_3d.py + _3d_common.py (#1, #8b, #8e, G1)

- #1: Split local_3d.py into text_to_3d.py (ShapETextTo3D), image_to_3d.py
  (ShapEImageTo3D, Hunyuan3D, StableFast3D, TripoSR, Trellis2, TripoSG),
  and _3d_common.py (shared helpers)
- #8b: Added env markers for triposg deps (rembg, diso Darwin exclusion),
  added shape and all-3d-pypi convenience groups
- #8e: Added SUPPORTED_PLATFORMS and INSTALL_HINT ClassVars to all nodes
- G1: Updated package_metadata JSON namespaces to match new modules
- Updated smoke tests to import from new modules
- Updated triposg/UPSTREAM.md reference

Agent-Logs-Url: https://github.com/nodetool-ai/nodetool-huggingface/sessions/c55f45dd-d177-4448-b960-536549c0b1ac


* Soften CUDA gates (#8i) and remove strong refs from node instances (#6, G6)

- #8i: SF3D and TripoSR emit warnings instead of raising RuntimeError
  on non-CUDA devices. Trellis2 and TripoSG remain CUDA-only.
- #6/G6: All heavy 3D nodes (Hunyuan3D, SF3D, TripoSR, Trellis2, TripoSG)
  no longer hold self._pipeline/self._model strong refs. Models are
  stored exclusively in ModelManager and fetched by cache key in process().
- Use _resolve_device() consistently across all nodes.

Agent-Logs-Url: https://github.com/nodetool-ai/nodetool-huggingface/sessions/c55f45dd-d177-4448-b960-536549c0b1ac


* Address code review: clarify ShapE base-class pattern, mark tasks done in FIX-MODEL3D.md

Agent-Logs-Url: https://github.com/nodetool-ai/nodetool-huggingface/sessions/c55f45dd-d177-4448-b960-536549c0b1ac


---------
